# docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  models.GeoPoint:
    properties:
      coordinates:
        items:
          type: number
        type: array
      type:
        type: string
    type: object
  models.Offer:
    properties:
      brand:
        description: 'Optional: Parsed brand like "Zara", "Starbucks"'
        type: string
      createdAt:
        description: When this offer was received
        type: string
      ctas:
        description: Call-to-action links
        items:
          type: string
        type: array
      id:
        type: string
      images:
        description: Images from the email
        items:
          type: string
        type: array
      sender_email:
        description: Email of the user who forwarded it
        type: string
      source:
        description: e.g., "email"
        type: string
      subject:
        description: Subject line of the email
        type: string
      tags:
        description: 'Optional: e.g., ["discount", "clothing"]'
        items:
          type: string
        type: array
      text:
        description: Plain text body
        type: string
    type: object
...
  models.Shop:
    properties:
      address:
        type: string
      brand:
        type: string
      category:
        description: e.g., "coffee", "clothing", "grocery"
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      location:
        $ref: '#/definitions/models.GeoPoint'
      name:
        type: string
      points_value:
        description: Points awarded for visiting
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API for the LoyalTea mobile app
  title: LoyalTea API
  version: "1.0"
paths:
  /offers:
    get:
      description: Retrieves a list of all available offers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Offer'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all offers
      tags:
      - offers
  /offers/{id}:
    get:
      description: Retrieves a single offer by its ID
      parameters:
      - description: Offer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Offer'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get offer by ID
      tags:
      - offers
...
  /shops:
    get:
      description: Get all shops or search by category or query string
      parameters:
      - description: Shop category
        in: query
        name: category
        type: string
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get shops
      tags:
      - Shops
    post:
      consumes:
      - application/json
      description: Create a new shop entry. Admin access only.
      parameters:
      - description: Shop data
        in: body
        name: shop
        required: true
        schema:
          $ref: '#/definitions/models.Shop'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Shop'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create a new shop (admin)
      tags:
      - Shops
  /shops/{id}:
    delete:
      description: Delete a shop by ID. Admin access only.
      parameters:
      - description: Shop ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a shop (admin)
      tags:
      - Shops
    get:
      description: Get a single shop by its ID
      parameters:
      - description: Shop ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Shop'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get shop by ID
      tags:
      - Shops
    put:
      consumes:
      - application/json
      description: Update an existing shop by ID. Admin access only.
      parameters:
      - description: Shop ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated shop data
        in: body
        name: shop
        required: true
        schema:
          $ref: '#/definitions/models.Shop'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Shop'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update shop (admin)
      tags:
      - Shops
  /shops/brand/{brand}:
    get:
      description: Get all shops that belong to a specific brand
      parameters:
      - description: Brand name
        in: path
        name: brand
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get shops by brand
      tags:
      - Shops
...
  /shops/nearby:
    get:
      description: Retrieve shops within a radius of the given latitude and longitude
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lng
        required: true
        type: number
      - description: 'Search radius in meters (default: 1000)'
        in: query
        name: radius
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get nearby shops
      tags:
      - Shops
...
schemes:
- http
swagger: "2.0"